# api/controlplane.openapi.yaml
openapi: 3.1.0
info:
  title: MicroProxy Control Plane API
  version: 0.1.0
  description: Control-plane contract for MicroProxy resources.
servers:
  - url: http://localhost:8081
paths:
...
  /api/v1/providers:
    get:
      summary: List providers
      operationId: listProviders
      responses:
        '200':
          description: Provider collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderListResponse'
    post:
      summary: Create provider
      operationId: createProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderWriteRequest'
      responses:
        '201':
          description: Provider created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderResponse'
        '409':
          description: Provider ID already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Semantic validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/providers/{providerID}:
    get:
      summary: Get provider
      operationId: getProvider
      parameters:
        - $ref: '#/components/parameters/providerID'
      responses:
        '200':
          description: Provider resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderResponse'
        '404':
          description: Provider not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      summary: Replace provider
      operationId: replaceProvider
      parameters:
        - $ref: '#/components/parameters/providerID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderWriteRequest'
      responses:
        '200':
          description: Provider replaced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderResponse'
        '404':
          description: Provider not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource version conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Semantic validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      summary: Patch provider
      operationId: patchProvider
      parameters:
        - $ref: '#/components/parameters/providerID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderPatchRequest'
      responses:
        '200':
          description: Provider patched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderResponse'
        '404':
          description: Provider not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource version conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Semantic validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      summary: Delete provider
      operationId: deleteProvider
      parameters:
        - $ref: '#/components/parameters/providerID'
        - name: resourceVersion
          in: query
          required: false
          schema:
            type: string
          description: Expected resource version. Can also be supplied in `If-Match` header.
      responses:
        '204':
          description: Provider deleted.
        '404':
          description: Provider not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource version conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Semantic validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    providerID:
      name: providerID
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ProviderSpec:
      type: object
      required: [id, name, type, endpoint]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        endpoint:
          type: string
    Provider:
      type: object
      required: [id, resourceVersion, spec]
      properties:
        id:
          type: string
        resourceVersion:
          type: string
        spec:
          $ref: '#/components/schemas/ProviderSpec'
    ProviderWriteRequest:
      type: object
      required: [provider]
      properties:
        resourceVersion:
          type: string
          description: Required for PUT and ignored for POST.
        provider:
          $ref: '#/components/schemas/ProviderSpec'
    ProviderPatchRequest:
      type: object
      required: [resourceVersion, patch]
      properties:
        resourceVersion:
          type: string
        patch:
          type: object
          minProperties: 1
          additionalProperties: false
          properties:
            name:
              type: string
            type:
              type: string
            endpoint:
              type: string
    ProviderResponse:
      type: object
      required: [provider]
      properties:
        provider:
          $ref: '#/components/schemas/Provider'
    ProviderListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Provider'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorModel'
    ErrorModel:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: validation_failed
        message:
          type: string
        request_id:
          type: string